backend/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.1
click==8.2.1
coverage==7.10.4
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.5.3
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
msgpack==1.1.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
pyflakes==3.3.2
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.35.0
